Create kali-setup.sh
This Bash script automates the process of updating and installing various software packages on a Kali Linux system. It includes tools like Vim, VirtualBox, Sublime Text, Gobuster, Wfuzz, Google Chrome, OpenVPN, and more. The script first updates the system, installs the specified packages, and then performs system upgrades. After the installations and upgrades are complete, the user has the option to restart the system. This script streamlines the initial setup of a Kali Linux system by handling installations and updates in a user-friendly manner.